+++ app/src/main/java/org/simple/clinic/patientattribute/entry/BMIEntryEffectHandler.kt
@@ -1,22 +1,14 @@
 package org.simple.clinic.patientattribute.entry
 
-import com.spotify.mobius.functions.Consumer
 import com.spotify.mobius.rx2.RxMobius
-import dagger.Lazy
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
 import io.reactivex.ObservableTransformer
-import org.simple.clinic.patientattribute.PatientAttributeRepository
-import org.simple.clinic.user.User
 import org.simple.clinic.util.scheduler.SchedulersProvider
-import org.simple.clinic.uuid.UuidGenerator
 
 class BMIEntryEffectHandler @AssistedInject constructor(
     @Assisted private val ui: BMIEntryUi,
-    private val patientAttributeRepository: PatientAttributeRepository,
-    private val currentUser: Lazy<User>,
-    private val uuidGenerator: UuidGenerator,
     private val schedulersProvider: SchedulersProvider,
 ) {
 
@@ -28,26 +20,13 @@ class BMIEntryEffectHandler @AssistedInject constructor(
   fun build(): ObservableTransformer<BMIEntryEffect, BMIEntryEvent> {
     return RxMobius
         .subtypeEffectHandler<BMIEntryEffect, BMIEntryEvent>()
-        .addTransformer(CreateNewBMIEntry::class.java, createNewBMIEntry())
-        .addAction(CloseSheet::class.java, ui::closeSheet, schedulersProvider.ui())
+        .addConsumer(
+            CloseSheet::class.java,
+            { effect -> ui.closeSheet(effect.bmiReading) },
+            schedulersProvider.ui()
+        )
         .addAction(ChangeFocusToHeight::class.java, ui::changeFocusToHeight, schedulersProvider.ui())
         .addAction(ChangeFocusToWeight::class.java, ui::changeFocusToWeight, schedulersProvider.ui())
         .build()
-  }
-
-  private fun createNewBMIEntry(): ObservableTransformer<CreateNewBMIEntry, BMIEntryEvent> {
-    return ObservableTransformer { createNewBMIEntries ->
-      createNewBMIEntries
-          .observeOn(schedulersProvider.io())
-          .map { createNewBMIEntry ->
-            patientAttributeRepository.save(
-                bmiReading = createNewBMIEntry.reading,
-                patientUuid = createNewBMIEntry.patientUUID,
-                loggedInUserUuid = currentUser.get().uuid,
-                uuid = uuidGenerator.v4(),
-            )
-            BMISaved
-          }
-    }
   }
 }