+++ app/src/main/java/org/simple/clinic/patientattribute/entry/BMIEntryUi.kt
@@ -1,7 +1,11 @@
 package org.simple.clinic.patientattribute.entry
 
+import org.simple.clinic.patientattribute.BMIReading
+
 interface BMIEntryUi {
-  fun closeSheet()
+  fun updateHeight(height: String)
+  fun updateWeight(weight: String)
+  fun closeSheet(bmiReading: BMIReading?)
   fun changeFocusToHeight()
   fun changeFocusToWeight()
   fun showBMI(bmi: String)