+++ app/src/main/java/org/simple/clinic/summary/PatientSummaryUpdate.kt
@@ -109,10 +109,12 @@ class PatientSummaryUpdate(
       is StatinInfoLoaded -> statinInfoLoaded(event, model)
       is AddTobaccoUseClicked -> dispatch(ShowTobaccoStatusDialog)
       is TobaccoUseAnswered -> dispatch(UpdateTobaccoUse(model.patientUuid, event.isSmoker, event.isUsingSmokelessTobacco))
-      is BMIReadingAdded -> dispatch(CalculateNonLabBasedCVDRisk(model.patientSummaryProfile!!.patient))
-      is AddBMIClicked -> dispatch(OpenBMIEntrySheet(model.patientUuid))
+      is BMIReadingAdded -> dispatch(CreateNewBMIEntry(model.patientUuid, event.bmiReading))
+      is AddBMIClicked -> dispatch(LoadBMiReading(model.patientUuid))
       is AddCholesterolClicked -> dispatch(OpenCholesterolEntrySheet(model.patientUuid))
       CholesterolAdded -> dispatch(CalculateLabBasedCVDRisk(model.patientSummaryProfile!!.patient))
+      is BMIReadingLoaded -> dispatch(OpenBMIEntrySheet(event.bmiReading))
+      is BMISaved -> dispatch(CalculateNonLabBasedCVDRisk(model.patientSummaryProfile!!.patient))
     }
   }