+++ app/src/test/java/org/simple/clinic/patientattribute/entry/BMIEntryUpdateTest.kt
@@ -8,38 +8,24 @@ import com.spotify.mobius.test.UpdateSpec
 import com.spotify.mobius.test.UpdateSpec.assertThatNext
 import org.junit.Test
 import org.simple.clinic.patientattribute.BMIReading
-import java.util.UUID
 
 class BMIEntryUpdateTest {
 
-  private val patientUuid = UUID.fromString("720f6fd4-4c4e-406a-b221-881990a962d4")
-
-  private val defaultModel = BMIEntryModel.default(patientUuid)
+  private val defaultModel = BMIEntryModel.default(BMIReading(height = 177f, weight = 63f))
 
   private val update = BMIEntryUpdate()
 
   private val spec = UpdateSpec(update)
 
   @Test
-  fun `when the save button is clicked, then save the bmi`() {
+  fun `when the save button is clicked, then close the sheet and return bmi reading`() {
     val model = defaultModel.weightChanged("63").heightChanged("177")
     spec
         .given(model)
         .whenEvent(SaveClicked)
         .then(assertThatNext(
             hasNoModel(),
-            hasEffects(CreateNewBMIEntry(model.patientUUID, BMIReading(height = model.height.toFloat(), weight = model.weight.toFloat())))
-        ))
-  }
-
-  @Test
-  fun `when bmi is saved, then close the sheet`() {
-    spec
-        .given(defaultModel)
-        .whenEvent(BMISaved)
-        .then(assertThatNext(
-            hasNoModel(),
-            hasEffects(CloseSheet)
+            hasEffects(CloseSheet(BMIReading(height = model.height.toFloat(), weight = model.weight.toFloat())))
         ))
   }
 
@@ -93,11 +79,12 @@ class BMIEntryUpdateTest {
 
   @Test
   fun `when weight backspace is clicked and is empty, then update the model and change the focus to height`() {
+    val model = defaultModel.weightChanged("")
     spec
-        .given(defaultModel)
+        .given(model)
         .whenEvent(WeightBackspaceClicked)
         .then(assertThatNext(
-            hasModel(defaultModel.deleteWeightLastDigit()),
+            hasModel(model),
             hasEffects(ChangeFocusToHeight)
         ))
   }
@@ -109,7 +96,7 @@ class BMIEntryUpdateTest {
         .whenEvent(BackPressed)
         .then(assertThatNext(
             hasNoModel(),
-            hasEffects(CloseSheet)
+            hasEffects(CloseSheet())
         ))
   }
 }